Move call note step views into their own modules
