+++ src/views/CallNoteSteps/__tests__/CallQuality-test.js
@@ -1,0 +1,22 @@
+import { noop } from 'lodash';
+import React from 'react';
+
+import { CallQuality } from 'src/views/CallNoteSteps';
+
+
+describe('CallNoteSteps', () => {
+  describe('CallQuality', () => {
+    it('should render', () => {
+      const el = render(
+        <CallQuality
+          callQuality="Delays"
+          onChangeText={noop}
+          onBackPress={noop}
+          onNextPress={noop}
+        />
+      );
+
+      expect(el.toJSON()).toMatchSnapshot();
+    });
+  });
+});
