+++ src/views/CallNoteSteps/__tests__/Mood-test.js
@@ -1,0 +1,21 @@
+import { noop } from 'lodash';
+import React from 'react';
+
+import { Mood } from 'src/views/CallNoteSteps';
+
+
+describe('CallNoteSteps', () => {
+  describe('Mood', () => {
+    it('should render', () => {
+      const el = render(
+        <Mood
+          menteeState="Sad"
+          onSelectImage={noop}
+          onBackPress={noop}
+          onNextPress={noop}
+        />
+      );
+      expect(el.toJSON()).toMatchSnapshot();
+    });
+  });
+});
