+++ src/views/CallNoteSteps/index.js
@@ -1,222 +1,5 @@
-import { upperFirst, isUndefined } from 'lodash';
-import React, { PropTypes } from 'react';
-import { View, Image, TouchableWithoutFeedback, ScrollView } from 'react-native';
-import { Text, MultiLineTextInput, RadioList } from 'src/components';
-
-import { Section } from 'src/views/Activity';
-import { FormStep } from 'src/components';
-import { MOOD_IMAGES } from 'src/constants/callNote';
-import images from 'src/constants/images';
-import styles from './styles';
-
-
-const Reflections = ({
-  reflection = '',
-  onChangeText,
-  ...props,
-}) => (
-  <FormStep
-    paginationBackDisabled
-    paginationDisabled={reflection.length === 0}
-    title="Please share your reflections of the discussion"
-    {...props}
-  >
-    <MultiLineTextInput
-      value={reflection}
-      placeholder="Type your answer here"
-      onChangeText={text => onChangeText({ reflection: text })}
-    />
-  </FormStep>
-);
-
-Reflections.propTypes = {
-  reflection: PropTypes.string,
-  onChangeText: PropTypes.func.isRequired,
-};
-
-
-const Mood = ({
-  menteeState = '',
-  onSelectImage,
-  ...props,
-}) => (
-  <FormStep
-    paginationDisabled={menteeState.length === 0}
-    title="How was your Mentee today?"
-    {...props}
-  >
-    <View style={styles.list}>
-      {Object.keys(MOOD_IMAGES).map(key => (
-        <TouchableWithoutFeedback
-          key={key}
-          onPress={() => onSelectImage({ menteeState: key })}
-        >
-          <View style={styles.item}>
-            <Image
-              style={[styles.image, menteeState === key && styles.imageIsSelected]}
-              source={MOOD_IMAGES[key]}
-            >
-            { menteeState !== key && <View style={styles.imageHighlight} /> }
-            </Image>
-            <Text style={styles.label}>{upperFirst(key)}</Text>
-          </View>
-        </TouchableWithoutFeedback>
-      ))}
-    </View>
-  </FormStep>
-);
-
-Mood.propTypes = {
-  menteeState: PropTypes.string,
-  onSelectImage: PropTypes.func.isRequired,
-};
-
-
-const Completed = ({
-  objectiveAchieved,
-  objective,
-  color,
-  onSelectImage,
-  ...props,
-}) => (
-  <FormStep
-    paginationDisabled={isUndefined(objectiveAchieved)}
-    title="Did you achieve the objective?"
-    {...props}
-  >
-    <View style={styles.completedContainer}>
-      <View style={styles.yesNoContainer}>
-        <TouchableWithoutFeedback onPress={() => onSelectImage({ objectiveAchieved: true })}>
-          <Image
-            source={
-              objectiveAchieved ?
-              images.CALL_NOTES_COMPLETED_YES_SELECTED :
-              images.CALL_NOTES_COMPLETED_YES
-            }
-          />
-        </TouchableWithoutFeedback>
-        <TouchableWithoutFeedback onPress={() => onSelectImage({ objectiveAchieved: false })}>
-          <Image
-            source={
-              objectiveAchieved === false
-                ? images.CALL_NOTES_COMPLETED_NO_SELECTED
-                : images.CALL_NOTES_COMPLETED_NO
-            }
-          />
-        </TouchableWithoutFeedback>
-      </View>
-      <View style={styles.objectiveContainer}>
-        <ScrollView>
-          <Section
-            color={color}
-            icon={images.ACTIVITY_OBJECTIVE}
-            title="Objective"
-          >
-            {objective}
-          </Section>
-        </ScrollView>
-      </View>
-    </View>
-  </FormStep>
-);
-
-Completed.propTypes = {
-  objectiveAchieved: PropTypes.any,
-  onSelectImage: PropTypes.func.isRequired,
-  objective: PropTypes.string,
-  color: PropTypes.string,
-};
-
-
-const Rating = ({
-  activityHelpful = '',
-  onChangeText,
-  ...props,
-}) => {
-  const items = [
-    'Not at all',
-    'A little',
-    'Somewhat',
-    'Quite a bit',
-    'A lot',
-  ];
-  const initialSelectedIndex = items.indexOf(activityHelpful);
-
-  return (
-    <FormStep
-      paginationDisabled={activityHelpful.length === 0}
-      title="Was the activity helpful in facilitating your discussion?"
-      {...props}
-    >
-      <View style={styles.ratingContainer}>
-        <ScrollView>
-          <View>
-            <RadioList
-              items={items}
-              onIndexChanged={(item) => onChangeText({ activityHelpful: item.item })}
-              initialSelectedIndex={initialSelectedIndex === -1 ? initialSelectedIndex : void 0}
-            />
-          </View>
-        </ScrollView>
-      </View>
-    </FormStep>
-  );
-};
-
-Rating.propTypes = {
-  activityHelpful: PropTypes.string,
-  onChangeText: PropTypes.func.isRequired,
-};
-
-
-const CallQuality = ({
-  callQuality = '',
-  onChangeText,
-  ...props,
-}) => {
-  const items = [
-    'Excellent',
-    'Ok',
-    "Couldn't hear",
-    'Call dropped',
-    'Delays',
-  ];
-  const initialSelectedIndex = items.indexOf(callQuality);
-
-  return (
-    <FormStep
-      paginationDisabled={callQuality.length === 0}
-      title="Rate the call quality"
-      {...props}
-    >
-      <View style={styles.callQualityContainer}>
-        <Text style={styles.callQualityHintText}>
-          Please help us improve our service by letting us know what the call quality was like.
-        </Text>
-        <ScrollView>
-          <View>
-            <RadioList
-              items={items}
-              onIndexChanged={(item) => onChangeText({ callQuality: item.item })}
-              initialSelectedIndex={initialSelectedIndex === -1 ? initialSelectedIndex : void 0}
-            />
-          </View>
-        </ScrollView>
-      </View>
-    </FormStep>
-  );
-};
-
-CallQuality.propTypes = {
-  callQuality: PropTypes.string,
-  onChangeText: PropTypes.func.isRequired,
-};
-
-
-export {
-  Reflections,
-  Mood,
-  Completed,
-  Rating,
-  CallQuality,
-};
+export { default as Reflections } from './Reflections';
+export { default as Mood } from './Mood';
+export { default as Completed } from './Completed';
+export { default as Rating } from './Rating';
+export { default as CallQuality } from './CallQuality';
